services/sentinel-brain/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
# UPDATE: Remove the specific version to get the latest (0.7.0+)
google-generativeai>=0.7.0
python-dotenv==1.0.1